+++ NEWS.md
@@ -56,6 +56,9 @@ Mutation Profiling:
   
 Distance Calculation:
 
++ Fixed a bug in `distToNearest` wherein `normalize="length"` for 5-mer models
+  was resulting in distances normalized by junction length squared instead of
+  raw junction length.
 + Added `findThreshold` function to infer clonal distance threshold from 
   nearest neighbor distances returned by `distToNearest`.
 + Renamed the `length` option for the `normalize` argument of `distToNearest`